autoMK early prototype
autoMK is functionally capable of replicating mouse and keyboard actions faithfully. It just entirely lacks all form of polish and a way to operate the program from outside an editor. You're looking at quite a few hours of work Jordan, if I charged even just $20 an hour for this then this Christmas present is kind of costly already. Your present better be equivalently valued!