autoMK giving up on function_list -> syntax
It's too hard to make the program perform exactly the way I envision it. In the future I might come back to it if Jordan really wants it in the code, but it's just damn difficult without relying on a bunch of arbitrary cutoffs e.g., how long does a key need to be pressed for for it to no longer be typing?